make SteamCmdParameters Builder an inner static class
