+++ src/main/java/cz/forgottenempire/arma3servergui/steamcmd/SteamCmdWorkerService.java
@@ -1,0 +1,5 @@
+package cz.forgottenempire.arma3servergui.steamcmd;
+
+public interface SteamCmdWorkerService {
+
+}
